♻️ Refactor_Extract countHour function
