+++ lib/pages/profile/dealHistory.dart
@@ -1,6 +1,7 @@
 import 'package:brainstorm_meokjang/models/deal.dart';
 import 'package:brainstorm_meokjang/utilities/Colors.dart';
 import 'package:brainstorm_meokjang/utilities/Popups.dart';
+import 'package:brainstorm_meokjang/utilities/count_hour.dart';
 import 'package:brainstorm_meokjang/utilities/domain.dart';
 import 'package:brainstorm_meokjang/utilities/rule.dart';
 import 'package:brainstorm_meokjang/widgets/all.dart';
@@ -103,25 +104,6 @@ class MyPostUnit extends StatefulWidget {
 }
 
 class _MyPostUnitState extends State<MyPostUnit> {
-  String countHour(DateTime givenDate) {
-    DateTime currentDate = DateTime.now();
-
-    Duration difference = currentDate.difference(givenDate);
-    int minutesDifference = difference.inMinutes;
-    int hoursDifference = difference.inHours;
-    int daysDifference = difference.inDays;
-
-    if (daysDifference >= 1) {
-      return '$daysDifference일 전';
-    } else if (hoursDifference >= 1) {
-      return '$hoursDifference시간 전';
-    } else if (minutesDifference == 0) {
-      return '방금 전';
-    } else {
-      return '$minutesDifference분 전';
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     return Padding(
