+++ lib/widgets/go_to_post/go_to_post_widgets.dart
@@ -1,26 +1,8 @@
 import 'package:brainstorm_meokjang/models/deal.dart';
 import 'package:brainstorm_meokjang/utilities/colors.dart';
+import 'package:brainstorm_meokjang/utilities/count_hour.dart';
 import 'package:brainstorm_meokjang/utilities/rule.dart';
 import 'package:flutter/material.dart';
-
-String countHour(DateTime givenDate) {
-  DateTime currentDate = DateTime.now();
-
-  Duration difference = currentDate.difference(givenDate);
-  int minutesDifference = difference.inMinutes;
-  int hoursDifference = difference.inHours;
-  int daysDifference = difference.inDays;
-
-  if (daysDifference >= 1) {
-    return '$daysDifference일 전';
-  } else if (hoursDifference >= 1) {
-    return '$hoursDifference시간 전';
-  } else if (minutesDifference == 0) {
-    return '방금 전';
-  } else {
-    return '$minutesDifference분 전';
-  }
-}
 
 class firstPostUnit extends StatefulWidget {
   final Deal deal;
